+++ src/feature/featurelistcallback.h
@@ -16,7 +16,7 @@
 #ifdef MZ_LINUX
 #  include <QProcessEnvironment>
 
-#  include "platforms/linux/linuxdependencies.h"
+#  include "platforms/linux/linuxutils.h"
 #  include "versionutils.h"
 #endif
 
@@ -95,7 +95,7 @@ bool FeatureCallback_splitTunnel() {
 
   /* Control groups v2 must be mounted for app/traffic classification
    */
-  if (LinuxDependencies::findCgroup2Path().isNull()) {
+  if (LinuxUtils::findCgroup2Path().isNull()) {
     return false;
   }
 
@@ -107,7 +107,7 @@ bool FeatureCallback_splitTunnel() {
   }
   QStringList desktop = pe.value("XDG_CURRENT_DESKTOP").split(":");
   if (desktop.contains("GNOME")) {
-    QString shellVersion = LinuxDependencies::gnomeShellVersion();
+    QString shellVersion = LinuxUtils::gnomeShellVersion();
     if (shellVersion.isNull()) {
       return false;
     }
@@ -115,7 +115,7 @@ bool FeatureCallback_splitTunnel() {
       return false;
     }
   } else if (desktop.contains("KDE")) {
-    QString kdeVersion = LinuxDependencies::kdeFrameworkVersion();
+    QString kdeVersion = LinuxUtils::kdeFrameworkVersion();
     if (kdeVersion.isNull()) {
       return false;
     }
