+++ src/platforms/linux/linuxdependencies.cpp
@@ -4,8 +4,6 @@
 
 #include "linuxdependencies.h"
 
-#include <mntent.h>
-
 #include <QDir>
 #include <QFile>
 #include <QFileInfo>
@@ -17,14 +15,6 @@
 namespace {
 
 Logger logger("LinuxDependencies");
-
-void showAlert(const QString& message) {
-  logger.debug() << "Show alert:" << message;
-
-  QMessageBox alert;
-  alert.setText(message);
-  alert.exec();
-}
 
 bool checkDaemonVersion() {
   logger.debug() << "Check Daemon Version";
@@ -77,82 +67,4 @@ bool LinuxDependencies::checkDependencies() {
   }
 
   return true;
-}
-
-// static
-QString LinuxDependencies::findCgroupPath(const QString& type) {
-  struct mntent entry;
-  char buf[PATH_MAX];
-
-  FILE* fp = fopen("/etc/mtab", "r");
-  if (fp == NULL) {
-    return QString();
-  }
-
-  while (getmntent_r(fp, &entry, buf, sizeof(buf)) != NULL) {
-    if (strcmp(entry.mnt_type, "cgroup") != 0) {
-      continue;
-    }
-    if (hasmntopt(&entry, type.toLocal8Bit().constData()) != NULL) {
-      fclose(fp);
-      return QString(entry.mnt_dir);
-    }
-  }
-  fclose(fp);
-
-  return QString();
-}
-
-// static
-QString LinuxDependencies::findCgroup2Path() {
-  struct mntent entry;
-  char buf[PATH_MAX];
-
-  FILE* fp = fopen("/etc/mtab", "r");
-  if (fp == NULL) {
-    return QString();
-  }
-
-  while (getmntent_r(fp, &entry, buf, sizeof(buf)) != NULL) {
-    if (strcmp(entry.mnt_type, "cgroup2") != 0) {
-      continue;
-    }
-    return QString(entry.mnt_dir);
-  }
-  fclose(fp);
-
-  return QString();
-}
-
-// static
-QString LinuxDependencies::gnomeShellVersion() {
-  QDBusInterface iface("org.gnome.Shell", "/org/gnome/Shell",
-                       "org.gnome.Shell");
-  if (!iface.isValid()) {
-    return QString();
-  }
-
-  QVariant shellVersion = iface.property("ShellVersion");
-  if (!shellVersion.isValid()) {
-    return QString();
-  }
-  return shellVersion.toString();
-}
-
-// static
-QString LinuxDependencies::kdeFrameworkVersion() {
-  QProcess proc;
-  proc.start("kf5-config", QStringList{"--version"}, QIODeviceBase::ReadOnly);
-  if (!proc.waitForFinished()) {
-    return QString();
-  }
-
-  QByteArray result = proc.readAllStandardOutput();
-  for (const QByteArray& line : result.split('\n')) {
-    if (line.startsWith("KDE Frameworks: ")) {
-      return QString::fromUtf8(line.last(line.size() - 16));
-    }
-  }
-
-  return QString();
 }
