+++ src/platforms/linux/linuxdependencies.h
@@ -10,10 +10,6 @@
 class LinuxDependencies final {
  public:
   static bool checkDependencies();
-  static QString findCgroupPath(const QString& type);
-  static QString findCgroup2Path();
-  static QString gnomeShellVersion();
-  static QString kdeFrameworkVersion();
 
  private:
   LinuxDependencies() = default;
